requirements.txt
# ✅ Flask & Backend Dependencies
Flask==3.0.2
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
gunicorn==21.2.0
fastapi==0.100.1
uvicorn==0.29.0
requests==2.31.0

# ✅ Database & ORM
mysql-connector-python==8.3.0

# ✅ Machine Learning & AI Models
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
scipy==1.13.0
tensorflow==2.15.0  # ✅ Works with Flower
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# ✅ Federated Learning (Flower) - 🚀 Easier than TFF!
flwr==1.8.0  # ✅ Using Flower instead of TensorFlow Federated

# ✅ Graph-Based Recommendations
networkx==2.8.8
node2vec==0.4.6
gensim==4.3.3

# ✅ Explainability & Visualization
shap==0.43.0
matplotlib==3.8.3
seaborn==0.13.2

# ✅ Utilities & Performance Optimization
joblib==1.4.2
tqdm==4.66.2
smart-open==7.1.0

